copy i2c_master_tx from system/i2c into skeleton/i2c, remove system/i2c
system/i2c contains working i2c master code
skeleton/i2c contains working i2c slave code
